merge1

+++ aten/src/ATen/core/blob.h
@@ -51,14 +51,14 @@ class CAFFE2_API Blob final : public c10::intrusive_ptr_target {
   /**
    * Returns the meta info of the blob.
    */
-  inline const TypeMeta& meta() const noexcept {
+  const TypeMeta& meta() const noexcept {
     return meta_;
   }
 
   /**
    * Returns a printable typename of the blob.
    */
-  inline const char* TypeName() const noexcept {
+  c10::string_view TypeName() const noexcept {
     return meta_.name();
   }
 
@@ -158,7 +158,7 @@ class CAFFE2_API Blob final : public c10::intrusive_ptr_target {
   void* ShareExternal(void* allocated, const TypeMeta& meta) {
     free_();
     meta_ = meta;
-    pointer_ = static_cast<void*>(allocated);
+    pointer_ = allocated;
     has_ownership_ = false;
     return allocated;
   }
@@ -166,7 +166,7 @@ class CAFFE2_API Blob final : public c10::intrusive_ptr_target {
   /**
    * Resets the Blob to an empty one.
    */
-  inline void Reset() {
+  void Reset() {
     free_();
     pointer_ = nullptr;
     meta_ = TypeMeta();
@@ -185,15 +185,14 @@ class CAFFE2_API Blob final : public c10::intrusive_ptr_target {
 
  private:
   void free_() {
-    if (has_ownership_) {
-      AT_ASSERTM(pointer_ != nullptr, "Can't have ownership of nullptr");
+    if (has_ownership_ && pointer_ != nullptr) {
       (*meta_.deleteFn())(pointer_);
     }
   }
 
   TypeMeta meta_;
-  void* pointer_ = nullptr;
-  bool has_ownership_ = false;
+  void* pointer_;
+  bool has_ownership_;
 
   C10_DISABLE_COPY_AND_ASSIGN(Blob);
 };
